requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
motor==3.3.2
pymongo==4.6.0

# Jira Integration
jira==3.5.2

# Git Integration
PyGithub==1.59.1
requests==2.31.0

# AWS Bedrock
boto3==1.34.0
botocore==1.34.0

# Microsoft Teams Bot
botbuilder-core==4.15.0
botbuilder-schema==4.15.0
botbuilder-integration-aiohttp==4.15.0

# LLM and AI
openai==1.3.7
langchain==0.0.350
langchain-community==0.0.10

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Code Intelligence
gitpython==3.1.40
